+++ Sources/Sentry/SentryRateLimitCategoryMapper.m
@@ -3,26 +3,37 @@
 #import "SentryRateLimitCategory.h"
 #import "SentryEnvelopeItemType.h"
 
+NS_ASSUME_NONNULL_BEGIN
+
 @interface SentryRateLimitCategoryMapper ()
 
 @end
 
 @implementation SentryRateLimitCategoryMapper
 
-+ (NSString *_Nonnull)mapEventTypeToCategory:(NSString *)eventType {
++ (NSString *)mapEventTypeToCategory:(NSString *)eventType {
     // Currently we classify every event type as error.
     // This is going to change in the future.
     return SentryRateLimitCategoryError;
 }
 
-+ (NSString *_Nonnull)mapEnvelopeItemTypeToCategory:(NSString *)itemType {
++ (NSString *)mapEnvelopeItemTypeToCategory:(NSString *)itemType {
+    NSString *category = SentryRateLimitCategoryDefault;
     if ([itemType isEqualToString:SentryEnvelopeItemTypeEvent]) {
-        return SentryRateLimitCategoryError;
+        category = SentryRateLimitCategoryError;
     }
     if ([itemType isEqualToString:SentryEnvelopeItemTypeSession]) {
-        return SentryRateLimitCategorySession;
+        category = SentryRateLimitCategorySession;
     }
-    return itemType;
+    if ([itemType isEqualToString:SentryEnvelopeItemTypeTransaction]) {
+        category = SentryRateLimitCategoryTransaction;
+    }
+    if ([itemType isEqualToString:SentryEnvelopeItemTypeAttachment]) {
+        category = SentryRateLimitCategoryAttachment;
+    }
+    return category;
 }
 
 @end
+
+NS_ASSUME_NONNULL_END