+++ Sources/Sentry/include/SentryRateLimitCategory.h
@@ -1,5 +1,8 @@
 #import <Foundation/Foundation.h>
 
+// more categories exist. New categories will be added and old SDKs are expected to fallback to `default`
+static NSString * const SentryRateLimitCategoryDefault = @"default";
 static NSString * const SentryRateLimitCategoryError = @"error";
 static NSString * const SentryRateLimitCategorySession = @"session";
-// more categories exist, but we only use error and session currently.
+static NSString * const SentryRateLimitCategoryTransaction = @"transaction";
+static NSString * const SentryRateLimitCategoryAttachment = @"attachment";